requirements.txt
# Python packages needed to be installed for functionality with API

Flask
flask-core
requests 
flask_cors
python-dotenv

# Can be installed with -> C: pip install -r requirements.txt